requirements.txt
streamlit==1.32.2
pandas==2.2.1
gspread==6.1.0
google-auth==2.28.2